getData bugfix and onDependencyChanged feature
getData bombs out on an item list with empty elements.  I added several tests for "s" before additional tests on properties of "s".  This allows dependency changes to continue even on item lists with empty elements.

onDependencyChanged, I passed the source variable from the function input vars to the schema resolver.  This allows the resolver to calculate which of an item list is making the modification.